chore: changes dir structure on machine learning dir
