--- machine_learning/trainer/data/raw_test_data.csv
@@ -1,120 +1,0 @@
-﻿temperature ,CO2,Light,Water,Sleep,current ELO,opponent ELO,Win black ,win white,Draw 
-19,1555,1576,4,454,2361,2158,1,0,0
-18,1372,1140,0,306,2672,1863,1,0,0
-22,852,1813,0,302,2709,1781,1,0,0
-19,1483,1671,4,541,1793,2338,0,0,1
-18,2000,1833,1,318,2073,2504,0,0,1
-20,936,1440,2,555,2400,2542,0,0,1
-22,1205,1662,2,424,1998,2023,1,0,0
-20,1150,1478,3,571,2684,1985,1,0,0
-19,1129,1505,3,355,2397,2021,0,1,0
-20,1081,1900,1,255,2629,2356,0,1,0
-19,1823,1203,4,497,2487,2414,1,0,0
-21,1105,1498,4,575,1873,2388,1,0,0
-18,1604,1444,1,483,2049,1993,0,1,0
-18,1643,1604,2,262,2046,1977,1,0,0
-20,1634,1094,4,357,2504,2737,1,0,0
-20,1653,1840,1,254,2382,2259,0,0,1
-21,1792,1084,3,495,2734,1899,1,0,0
-21,1158,1642,4,507,1986,2125,0,1,0
-18,1770,1229,3,347,2391,2006,0,1,0
-19,1761,1613,0,472,1924,2743,0,1,0
-20,1236,1111,0,561,2106,2567,0,1,0
-22,987,1089,4,272,1989,2334,0,1,0
-19,1877,1171,4,310,2650,2449,0,1,0
-18,1366,1595,1,336,2007,2509,0,1,0
-21,1107,1701,2,316,2656,2571,0,0,1
-19,1910,1311,1,569,2632,1986,1,0,0
-20,956,1442,0,530,2555,2546,1,0,0
-21,1947,1023,2,400,2657,2635,0,1,0
-20,854,1946,0,513,1908,2461,1,0,0
-19,1898,1501,1,425,2191,1954,1,0,0
-22,1429,1942,3,363,2146,2390,1,0,0
-22,831,1666,4,264,1986,2025,0,0,1
-18,1627,1168,4,306,2103,2655,0,1,0
-18,1503,1118,3,456,1760,1778,0,1,0
-22,1680,1606,4,595,2526,1877,1,0,0
-18,832,1881,3,528,2093,2448,0,1,0
-21,1831,1858,3,391,1769,2646,1,0,0
-19,1230,1036,1,367,2073,2135,0,1,0
-22,1546,1854,0,422,2362,2317,1,0,0
-20,1364,1479,1,410,2376,1930,0,0,1
-20,1205,1228,4,397,2636,2441,1,0,0
-22,1017,1252,1,295,2360,2015,1,0,0
-18,1584,1849,0,541,2213,2474,1,0,0
-22,1605,1537,4,309,2160,1803,0,1,0
-19,1704,1127,3,535,1809,2629,1,0,0
-22,965,1840,4,274,2221,2133,1,0,0
-19,898,1440,2,415,2416,2615,1,0,0
-21,983,1351,3,567,2339,1986,0,0,1
-21,1583,1320,0,318,2017,2011,1,1,0
-22,892,1495,0,324,2534,1961,0,1,0
-20,1479,1353,4,527,2243,2381,1,0,0
-18,1596,1147,1,495,1895,2184,0,1,0
-21,1332,1691,0,478,2306,2641,1,0,0
-20,1725,1208,3,393,2118,2524,0,0,1
-21,1919,1597,4,333,2112,1869,0,0,1
-18,1489,1829,0,373,2593,2051,0,0,1
-21,1788,1479,4,281,1752,2362,0,1,0
-22,980,1595,2,586,2117,2210,0,0,1
-21,921,1398,4,437,2652,2594,1,0,0
-22,1350,1406,4,532,1968,1945,1,0,0
-21,1818,1670,1,591,1966,1804,0,1,0
-20,1056,1284,2,364,2431,2200,0,1,0
-18,1116,1039,1,363,1828,1927,1,0,0
-22,1973,1830,1,448,2107,2323,0,1,0
-18,1554,1736,3,404,2415,1980,1,0,0
-19,1520,1243,0,574,2613,1832,0,1,0
-20,1368,1461,0,288,2011,2463,1,0,0
-19,908,1044,3,288,2243,1754,1,0,0
-19,1569,1808,0,288,2619,2655,1,0,0
-22,1741,1976,1,420,2014,1860,1,0,0
-20,826,1073,4,510,2427,2047,1,0,0
-21,1817,1281,4,442,2499,1802,1,0,0
-19,1876,1745,2,502,2393,2526,0,0,1
-19,1206,1128,3,436,2057,1845,0,0,1
-18,1182,1566,4,351,2729,2344,1,0,0
-20,1672,1587,1,328,2194,2191,0,1,0
-19,803,1835,4,395,1899,2456,1,0,0
-22,1531,1008,4,555,1892,2161,1,0,0
-21,1587,1051,4,362,2068,2273,1,0,0
-22,1327,1333,0,573,2343,2388,0,1,0
-22,1974,1353,0,545,1881,2078,1,0,0
-19,820,1075,1,400,2103,2164,0,1,0
-18,945,1863,2,280,2663,2240,0,1,0
-22,1255,1981,2,573,2594,2707,0,1,0
-20,1797,1979,1,599,1780,2028,0,1,0
-20,1036,1429,0,329,1997,2725,1,0,0
-22,1031,1605,2,260,2738,2411,0,0,1
-21,872,1447,2,554,1879,2720,0,1,0
-18,1360,1970,4,498,2359,2704,0,1,0
-20,1861,1689,2,254,2591,2601,1,0,0
-22,1997,1211,4,486,2292,2126,1,0,0
-21,1067,1262,3,297,2592,2468,1,0,0
-22,1185,1163,0,525,2506,2427,0,0,1
-19,1487,1716,3,522,1794,2727,1,0,0
-21,1634,1734,4,252,2572,2187,1,0,0
-19,1333,1004,3,268,2296,1823,0,1,0
-18,1582,1096,0,549,1940,2651,0,0,1
-22,1490,1039,1,304,2594,2195,1,0,0
-22,1023,1269,2,332,2566,2671,1,0,0
-22,1519,1674,1,474,2059,2628,0,1,0
-20,1942,1290,1,568,2218,2264,0,0,1
-20,1848,1658,3,498,2566,2584,1,0,0
-20,1028,1842,1,316,2431,2533,0,1,0
-20,920,1650,3,478,2061,1966,0,1,0
-22,1371,1555,3,427,2084,2733,0,1,0
-21,1843,1960,4,529,2314,1905,0,1,0
-22,1303,1625,2,257,2435,1827,0,1,0
-22,1095,1380,2,373,1839,2061,0,1,0
-22,1095,1982,1,559,2025,2599,1,0,0
-18,1916,1030,2,314,2330,2468,0,1,0
-18,1197,1281,2,260,1973,1932,0,0,1
-22,1878,1119,4,325,2490,2723,1,1,0
-22,1550,1174,2,289,2269,2504,0,0,1
-20,1020,1210,0,375,2216,1911,0,1,0
-18,1761,1838,0,570,2110,1774,0,1,0
-18,1047,1622,0,435,2279,2429,0,0,1
-22,977,1281,1,486,2272,2247,1,0,0
-21,1596,1545,4,479,2453,2049,1,0,0
-21,1424,1749,2,519,2703,1805,1,0,0
