+++ machine_learning/trainers/trainer-winrate/requirements.txt
@@ -1,0 +1,8 @@
+scikit-learn
+pandas
+numpy
+mlflow
+joblib
+uvicorn
+kagglehub
+seaborn
