+++ machine_learning/trainers/trainer-winrate/steps/4_evaluate.py
